backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
Pillow
pytesseract
python-multipart
pymupdf
supabase